storage/requirements.txt
azure-mgmt-resource
azure-identity
azure-storage-blob
